requirements.txt
langchain==0.1.17
langchain-community==0.0.37
langchain-core==0.1.52
langchain-openai==0.1.6
python-dotenv==1.0.1
streamlit==1.34.0
streamlit-float==0.3.4
sqlalchemy==2.0.30
